chore(refactor): move util/*reducers into format/ dir
